Reworked helpers
